requirements.txt
django==4.1.4
django-solo==2.0.0
django-ordered-model==3.6
django-environ==0.9.0
psycopg2-binary==2.9.5